Use React/Redux frontend, instead of AngularJS.

+++ sut/start.sh
@@ -10,17 +10,29 @@ sleep 5
 ## Test backend endpoint
 curl 'http://localhost:3000/api/tags'
 
-## Start app frontend
-cd $SUT_FOLDER/node_modules/conduit-angularjs/
+
+## Start AngularJS app frontend if requested
+if [[ "$START_ANGULARJS" ]]; then
+  cd $SUT_FOLDER/node_modules/conduit-angularjs/
+  yarn
+  cp $SUT_FOLDER/app.constants.js.modified ./src/js/config/app.constants.js
+  cat ./src/js/config/app.constants.js
+  rm -rf ./dist
+  ./node_modules/.bin/gulp build
+  test -e ./dist
+  yarn add http-server
+  ./node_modules/.bin/http-server dist/ -a localhost -p 4000 &
+  sleep 5
+
+  ## Test frontend endpoint
+  curl 'http://localhost:4000/'
+fi
+
+## Start React-Redux app forntend
+cd $SUT_FOLDER/node_modules/react-redux-realworld-example-app/
+patch --forward ./src/agent.js $SUT_FOLDER/agent.js.patch || true
 yarn
-cp $SUT_FOLDER/app.constants.js.modified ./src/js/config/app.constants.js
-cat ./src/js/config/app.constants.js
-rm -rf ./dist
-./node_modules/.bin/gulp build
-test -e ./dist
-yarn add http-server
-./node_modules/.bin/http-server dist/ -a localhost -p 4000 &
-sleep 5
+BROWSER=none yarn start &
 
 ## Test frontend endpoint
-curl 'http://localhost:4000/'
+curl 'http://localhost:4100/'
